pyproject.toml
[project]
name = "pa_call_center_agent"
version = "0.1.0"
description = "LangGraph-based PA Call Center Agent"
requires-python = ">=3.10"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "pydantic",
    "python-dotenv",
    "jinja2",
    "streamlit",
    "watchdog" 
]

[tool.setuptools]
packages = ["app", "utils", "schemas", "data"]
